Continued work on the Items window
Modified all files inside the Items folder and made minor changes for the clsDataAccess and clsItems files.

Added all the functions to the clsItemsLogic file, moved all the logic from the clsItemsSQL file into the clsItemsLogic file, motified the datagrid inside the wndItems.xaml so it displays the item's, and implemented  the majority of the event listeners, I think will be used.

Changed the clsDataAccess class so it connects to the correct DB and additional functionality to the clsItems file, majority of the changes is so that it implements the interface "INotifyPropertyChanged".