Started process of making the UI more functional.
Added two new custom user controls for to allow the user to add/update a item.  The buttons are not yet fully programmed, but they have the event delegates required to implement them.  Also made changes to the wndItems.xaml file to accompany this change of data inputs.

Also modified the App.xaml to include more styles that the application will use.  In addition, copied over the Theme.xaml file from the SearchWindow branch to the theme is applied to this branch.

Current error, the preview inside VS shows the background color but when you run it, it doesn't show the color, it defaults to white not DarkSlateGray.