99% finished making the UI more functional.
Remove the custom user controls for the items windows because they both act almost the same.  So I can have it be just the same set of controls and modify their properties to accomplish what I need them to do.  And as stated in the summary am about done with the items window, I still want to test all possible conditions it might come across but I am done working on if for today.

Added comment headers for all the major classes/windows for everyone to fill out when they have the time.