Finish prototype for Items window
Finished the prototype for the Items window.  It now is able to insert, update, and delete any item to/from the DB and displays the DB to the user in a DataGrid.

But there is still the error about the background color not being DarkSlateGray when it should be.  I will be looking into this to see why it is happening, but it would help if everyone could at least take a look at it.