Minor changes to how the Items window works
Performed a few modifications to the Items window, files attached to the Items window, and changed the method of how the Main window knows when if an item has been modified.  Instead of having a public variable, it returns a DialogResult of True or False depending on if item(s) has been modified or not.