Changed how cost is inputted and how the window resizes
Changed the cost field so it only accepts integers and made the window so it resizes it's width to fit it's content.  Also added the changes to the invoice and item classes so they make use of the correct data types.  The same thing goes for the search window (only so it doesn't cause errors when trying to build the program).